requirements.txt
fastapi
uvicorn
google-generativeai
google-api-python-client
python-dotenv